Bug 1850739 [wpt PR 41714] - Invalidation on stylesheet changes using selector matching, a=testonly
Automatic update from web-platform-tests
Invalidation on stylesheet changes using selector matching

When inserting, deleting or modifying stylesheets, we need to invalidate
style for some elements, even though the elements themselves did not
change. Currently, we reuse the concept of InvalidationSets, which are
mostly designed for changes to elements, for that. This can sometimes
cause overbroad invalidations, especially if the style is written in
a “modern” (read: less-than-ideal) fashion.

We replace this by a somewhat slower but more precise and probably
simpler system; for each candidate element in the DOM, we see if it
matches any selectors in the old or the new rulesets (by means of
the same bucketing logic we currently have during element rule
collection), and if so, we mark it for style recalc. This isn't perfect;
if we are e.g. adding a single rule to an existing stylesheet, we would
still invalidate based on all rules in the sheet. We hope to improve
this in a future patch, but even now, this shows significant
improvements over most of blink_perf.css, with only a few regressions.

Note that this means the experimental style invalidation information in
devtools becomes somewhat less useful; it will no longer give reasons
for invalidations when style sheets change. Fixing this would probably
involve extending the protocol, although I haven't looked deeply into
this.

Bug: 1444522
Change-Id: Ib95cf8d215bf4e7b58ab1955ebc1f3c7e31aead8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4783456
Commit-Queue: Steinar H Gunderson <sessechromium.org>
Reviewed-by: Adithya Srinivasan <adithyaschromium.org>
Reviewed-by: Rune Lillesveen <futharkchromium.org>
Cr-Commit-Position: refs/heads/main{#1190433}

--

wpt-commits: 1aaf432aa82f200f0aa06a9c34b9f494024e2831
wpt-pr: 41714

UltraBlame original commit: 1dbf941d1dff82e675f03c144f86e26da522850a