Bug 1852973 - Delegate the responsibility of opening about:addons from the migration wizard to MigrationWizardParent. r=tgiles
Originally, we had had the extensions migration success link work by using a normal
anchor element with "about:addons" set as the href. This works on privileged pages like
about:preferences, but not for unprivileged pages like about:welcome - unprivileged
pages are generally unable to directly link to most about: pages.

The solution here is to continue to use an anchor element, but to use an event handler
on it rather than a "href" attribute, and have the MigrationWizardParent hear that
click event and take responsibility for opening about:addons.

As part of this, we're also intentionally choosing to open about:addons in a background
tab so that the user can continue to work through their onboard flow uninterrupted
should they click on the link while on about:welcome.

Differential Revision: https://phabricator.services.mozilla.com/D188116

UltraBlame original commit: 41001d4d5975f984e65c4bd8349a939942f6f7a3