Bug 1853630 use last value of previous event for start of SetTargetAtTime() curves r=padenot
Previously this was happening only after CleanupEventsOlderThan() had been
called between blocks.

t == t0 means the exponential is 1 and the curve uses the value from the curve
of the previous event.
When the time constant is zero and t == t0, the exponential argument is
undefined, but v0 at t0 is the limit as the time constant approaches zero.

Differential Revision: https://phabricator.services.mozilla.com/D188437

UltraBlame original commit: 054d3b1a7f96127064da282d87420fadda4b7a53