Bug 1854168 - Handle when ShoppingProduct returns null from requestAnalysis. r=shopping-reviewers,jhirsch
If the analysis request returns null, there was an error while requesting the data so the UI should display the generic error message.
This adds a default error to the data if the analysis data is null to pass along to the shopping container.

Differential Revision: https://phabricator.services.mozilla.com/D188772

UltraBlame original commit: 1cd46e0b23dac1acd358b8e4687fd5ee656d9912