Bug 1812135 - Use moz-toggle in the about:newtab personalization panel. r=thecount,amy,hjones,desktop-theme-reviewers,bolsson
This adds moz-toggle.mjs as a module to load in the main newtab document
(cached or otherwise), and then converts the existing toggles to use
moz-toggle.

This also required me to bring in tokens-brand.css to get the right
colours for the toggle to appear.

There was some markup in the original panel that had the label exist outside
of the toggle, and some CSS grid styling to make it appear alongside it.
Thankfully, moz-toggle takes care of a lot of that for us, so I was able
to get rid of the external label element and the associated styles.

Differential Revision: https://phabricator.services.mozilla.com/D187716

UltraBlame original commit: 6a7f708c3d0458c9b30efa2bdfde6e7812554d1e