Bug 1850573 - Override number of style threads on systems with heterogeneous CPUs. r=emilio
If there are 2 or more "big" CPUs then use the number of big CPUs as
the number of threads. If there are fewer than 2 then use the number
of big plus the number of "medium".

Currently this only affects android, as hal::GetHeterogeneousCpuInfo
is not implemented on other platforms.

Differential Revision: https://phabricator.services.mozilla.com/D188480

UltraBlame original commit: 80bdce072953e3d7600b588635cba6fba85a8fd0