Bug 1834274 - Part 2: Implement the FingerprintingWebCompatService. r=tjr
This patch implements the FingerprintingWebCompatService as a
remote settings client. The service will talk to the remote setting
collection "fingerprinting-protection-overrides" and the local testing
pref "privacy.fingerprintingProtection.granularOverrides" for the
fingerprinting overrides. Then, it calls to the nsIRFPService to
set the overrides to it.

Differential Revision: https://phabricator.services.mozilla.com/D182810

UltraBlame original commit: ea9df179fa4cd022f8438b156e2531aef9f5de96