Bug 1855312 - Check against all possible 'Mach' sites when activating telemetry, rather than just the `mach` site r=sheehan
Changes in Bug 1695312 made it so the actual command site is activated
much sooner, as such, at this point in the program execution, any of the
'Mach' sites can be active, so we should check against all of them.

In the past, the 'mach' site was the only one that could be active at
this time, so it was the only one that was needed to be checked for.

Differential Revision: https://phabricator.services.mozilla.com/D189290

UltraBlame original commit: 72153d629a084254f7067b3eeb48181de01952f7