Bug 1851913 - Remove newtab's explicit meow dependency. r=omc-reviewers,negin
Since stylelint is now a dependency of the project root, and meow is
a dependency of stylelint, we don't need to explicitly list meow as a
dependency of newtab. The project root uses a newer version of meow that
requires some small workarounds to work outside of an ES module, so we
also need to update the newtab scripts that use meow.

Differential Revision: https://phabricator.services.mozilla.com/D187603

UltraBlame original commit: 6fc66d324cb8d3114020c6bfc3e66d7a8f3b5e45