Bug 1855705 part 2 - Add a cache for Object.assign's fast path. r=jonco
The fast path we have for plain objects only works with data properties stored
in the shape, so a cache based on the shapes of the two objects works really
well for this code.

Even though it's a one-entry cache, the `lookup` method has a hit rate of
at least 90% on both Speedometer 2 and 3 so it's very effective.

Differential Revision: https://phabricator.services.mozilla.com/D189504

UltraBlame original commit: 7728a6fdad9443e5803c4aa696e395a627b1d384