Bug 1855716 - Move `jsonschema`, `pyrsistent`, and `pkgutil_resolve_name` dependencies back into `mach.txt` instead of specifying them in individual sites r=mach-reviewers,ahal
D180501 made the original change that this patch undoes. An earlier
change in that stack inadvertently silently disabled Telemetry, which
hid this issue.

The problem is that `glean` (which we optionally install in the `mach`
site also needs `jsonschema` (and `jsonschema` depends on `pyrsistent`
and `pkgutil_resolve_name`). Since our vendored version was no longer
specified in the `mach.txt` site, it pulled a different version
(`4.19.1`) from the vendored version (`4.17.3`) which causes issues in
the `wpt` site.

Since glean explicitly needs `jsonschema` it never should have been
taken out of the `mach` site, so this patch corrects that mistake.

Differential Revision: https://phabricator.services.mozilla.com/D189553

UltraBlame original commit: 50021661148610424471894e9e774198108e1bf1