Bug 1853448 - part5 : make HEVC test run on all platforms, but only check playback on Windows. r=media-playback-reviewers,padenot
HEVC playback needs hardware decoding, and it currently only support on
Windows. HEVC playback check would be run when the task is in the
mda-gpu, which has the ability for hardware decoding. On other
platforms, HEVC should not be supported.

We also need to set the pref in the `ini` file, instead of setting it in
file. Because that pref would be checked while creating GPU process,
setting in file would be too late for us.

In addition, interesting thing for HEVC on media engine is that, the
media engine can even play HEVC on the machine which doesn't support
hardware decoding. I suspect that media engine still supports SW HEVC
decoding but just not expose that feature to other applications.

Differential Revision: https://phabricator.services.mozilla.com/D189030

UltraBlame original commit: b1b8fbac2eabba889167c0501ec01c95a646e6e4