Bug 1699328 - Avoid collisions with DbgHelp symbol sessions from other components when walking the stack on Windows. r=bobowen
SymInitialize can fail with ERROR_INVALID_PARAMETER if some other piece
of code has already called it with the handle value that we are
providing. Therefore it is not recommended to pass GetCurrentProcess()
to this function. Instead, this patch duplicates the handle for the
current process, so that we can pass a unique handle to the current
process and thus avoid collision with handle values that other
components might pass to SymInitialize.

Differential Revision: https://phabricator.services.mozilla.com/D189360

UltraBlame original commit: 161d8267260601a9d3eabc924298b5ba11a03d44