Bug 1855887 - Improve handling of keywords in substitute_references_in_value_and_apply. r=emilio
In bug 1840478, two follow-up TODO questions were left in
substitute_references_in_value_and_apply regarding handling of CSS
keywords such as 'unset'. For non-inherited custom properties, we don't
need to put any value in the map of computed properties when that value
is the initial one, so just remove it. For inherited custom properties,
we could probably do the same when that value is the inherited one and
postpone actually deep cloning the inherited map at the end of the
build method. Comment is updated to refer to bug 1855887 instead.

Differential Revision: https://phabricator.services.mozilla.com/D189740

UltraBlame original commit: 2f55472aa74398aac03d9fa42ccca57b7c667037