Bug 1840965 - Create a decoder template r=padenot
This patch creates a C++ template that will be used as a base class of
WebCodecs' VideoDecoder and (future) AudioDecoder.

Currently, only the internal member variabls and functions are moved
into the template class. The common interfaces shared between the
VideoDecoder and the AudioDecoder will be moved to the template in the
following patches.

Note that some changes are made in this patch:
- Add a VideoDecoderTraits type to specify the types used in the decoder
  template
- Reset(const nsresult& aResult) and Close(...) for internal usages are
  renamed to ResetInternal in order to be distinguished from the
  Reset(ErrorResult& aRv) in the VideoDecoder's public interface
- ControlMessage-related classes become nested classes since they have
  member variables whose types depends on the template's typename
- DecodedDataToVideoFrame is generalized to a function named
  DecodedDataToOutputType that convert the decoded data to a specific
  type (e.g., VideoFrame for VideoDecoder)
- Rename CreateVideoInfo to CreateTrackInfo and move it to
  VideoDecoderTraits so that the decoder template can get a TrackInfo
  from the template inheritance
- Add a IsSupported method to VideoDecoderTraits so that the decoder
  template can know what codec its template ingeritance supports

Differential Revision: https://phabricator.services.mozilla.com/D188538

UltraBlame original commit: 4c1d4aba133c2e407032dcd9bb340bec7dbfddae