Bug 1856289 - Make the test request a reflow of a child document when it received the last expected event message r=edgar
The test requests reflow to a child document which the listener listens to
when expected event gets received as a message.  However, the message listener
is set twice for listening same events on `<html>` and `<div>`.  Therefore,
the request occurs twice, and that assumes that the reflow happens before
the next task of a `setTimeout` call.

This patch makes the order stabler.
* Request reflow only when the the last expected event is received
* Remove the message listeners when the reflow request is performed in the child

Differential Revision: https://phabricator.services.mozilla.com/D189817

UltraBlame original commit: 0d3b6ffb57f5832883618dd928d1f080fece7874