Bug 1857394. Enable layout/reftests/bugs/1743533-1.html 1743533-2.html with fuzz. r=emilio
This should be good enough to catch regressions. The main thing is we paint something in the fragments of the iframe on the second and following pages and this should catch it if we fail to.

Fuzz amounts are from an all reftests try run.

Differential Revision: https://phabricator.services.mozilla.com/D190246

UltraBlame original commit: 601ee1dffd9780bd41e93ee188496c396d7ce235