Bug 1858041 - Temporarily fail-if some Tier 2 a11y_checks failing on Autoland. r=Jamie,devtools-reviewers,mixedpuppy,nchevobbe
We are working on [enabling Tier 2 a11y-checks to ensure our products have basic accessibility built-in (bug 1692110)](https://bugzilla.mozilla.org/show_bug.cgi?id=1692110) and before that, we need to prepare the existing code base. While we triage and investigate existent tests that would fail once the a11y-checks are enabled in the CI, we are going to expect to be failing these tests altogether and then handle each one of them individually.

This task is to expect failing of these tests with `fail-if` condition of the failing tests while we continue investigation into the reasons these tests failed and to be able to backtrack these tests once they're resolved.

Related bugs:
- Fail-if tests: meta bug 1848402
- Explore capturing click events for a11y_checks: bug 1692110

Differential Revision: https://phabricator.services.mozilla.com/D190524

UltraBlame original commit: 9dc8d4452ed5f04f24b67c70be042614f499ff48