Bug 1844092 - [devtools] Make Stacktrace frame location keyboard accessible. r=devtools-reviewers,bomsy.
The frame elements already had the correct roles assigned,
but this was made completely useless because we had a
display: contents on them, which strips any accessibility role.
This was done for layout purpose that we now can do with
subgrids, so we update the component's CSS.
A webconsole test is added to ensure those stactkrace frames
can be navigated with the keyboard.

Differential Revision: https://phabricator.services.mozilla.com/D190594

UltraBlame original commit: 1f3e7fe297e25355ebc02a5c0320f3f7e35de146