Bug 1859200 introduce MediaTrack::ControlMessageInterface r=pehrsons
Having this in MediaTrackGraph.h will enable some templated ControlMessage
methods on MediaTrack without needing to put their definitions in
MediaTrackGraphImpl.h.

Reference-count logging is on the abstract class because that will be the base
of a templated class in a subsequent comment and the logging doesn't like the
same name on classes of different sizes.

Differential Revision: https://phabricator.services.mozilla.com/D191060

UltraBlame original commit: 29f157aa1c57254a930c280fbe8ce12f5bfa9bc1