Bug 1854090 - Omit psnames of fonts with family-name *.tmp from src:local() lookups on Windows. r=jwatt
This is a sad hack, but aims to work around the issue that some
PDF-related software (Acrobat is suspected but not currently confirmed)
is potentially polluting the global font collection with re-encoded
subsets of standard fonts like Arial, but does not munge the psname;
these can then be returned by src:local(...) lookups, which results in
garbled or missing text.

In principle, if such fonts are "installed", Firefox is not wrong to use
them; it's just a badly-configured system. But given that it seems to be
caused by some PDF-handling software, without the user's knowledge, it
seems worth trying to avoid the problem. Simply skipping the psname of
any font with family-name *.tmp is highly unlikely to adversely affect
any real-world content.

Differential Revision: https://phabricator.services.mozilla.com/D191639

UltraBlame original commit: b2085aad8c83b38147e4a8b01886f825a9e0f838