Bug 1832059 - Add a test runner for functional tests to mozperftest. r=perftest-reviewers,kshampur
This patch adds a FunctionalTestRunner class that can be used by all the layers to run `./mach test` which lets us easily run all functional tests locally. It will be used for running mochitest tests, and in the future, it may replace parts of the xpcshell layer.

The runner can only be used locally, and only needs a test path provided. It will then run all the manifests that it finds that test in (but only running that single test). At the same time, it'll parse the logs produced by `./mach test` with FunctionalTestProcessor to find all the metrics that were output. However, it does not parse them into a JSON format, that's left for the layers to handle.

Differential Revision: https://phabricator.services.mozilla.com/D190993

UltraBlame original commit: eafb5db59a2c4c9d2416438cb20daa478493ee67