Bug 1860643 - Disable a11y_check for two mouse-specific test cases on .brand-logo in Firefox View. r=kcochrane,Jamie,fxview-reviewers
There are 2 failing test cases for the `div`/`img` with class name `brand-logo` that appear to test not user-facing behavior: the test cases are sending clicks to brand-logo as a container. This test target appears to be to confirm the feature tour pref change after a custom FEATURE_CALLOUT messages are dismissed by these clicks on a non-actionable element, while there are other ways to dismiss/pass the tour messages that are accessible.

This is not expected or does not have to be done by a user of an assistive technology and it does not have to confirm the accessibility of this element.

Thus, we are excluding these click events from accessibility checks and removing the `fail-if` annotation from the test manifest that was added by bug 1854625 and meta bug 1858037 before the investigation in the meta bug 1854233.

Differential Revision: https://phabricator.services.mozilla.com/D191686

UltraBlame original commit: 09e0e517398773d30ba186e6cbcab04b41ac1f4f