Bug 1839208 - Switch the newtab Redux JSM to an ESM. r=sclements,Mardak,sylvestre
I had to make a minor adjustment to the symbolObservablePonyfill thing inside
of the redux.js code because Symbol.observable is not something that can be
written to in the privileged module context. Hopefully the comment I added
in the code makes that clear!

Differential Revision: https://phabricator.services.mozilla.com/D191348

UltraBlame original commit: ff23efb9dc2c25ece6188efed6127af4faab7b27