Bug 1861253 - Draw select's menulist arrow in the select's color. r=dholbert
In the regressing bug 1849793 I removed a hunk like this because we
don't use any of the parent state on the parent properly:

```
  if (aAppearance == StyleAppearance::MozMenulistArrowButton) {
    // HTML select and XUL menulist dropdown buttons get state from the
    // parent.
    nsIFrame* parentFrame = aFrame->GetParent();
    aFrame = parentFrame;
    elementState = GetContentState(parentFrame, aAppearance);
  }
```

However that aFrame = parentFrame also changed the frame we read the
color for the fill. This is not an issue for menus because those don't
override the colors, but for the select menulist button we need to
explicitly inherit.

Differential Revision: https://phabricator.services.mozilla.com/D192002

UltraBlame original commit: dd29bee48c0c442e848435c8138bd6e38ae9303c