Bug 1523435 - Do not open two tabs for talos startup tests if the profiler is enabled r=perftest-reviewers,sparky
It looks like this is a regression from Bug 1121571. This was opening two tabs
instead of one because we are already adding the `url` to the command line
arguments below:
https://searchfox.org/mozilla-central/rev/40d51bef58b8e901d6ab4b60dd280f372a0e417d/testing/talos/talos/utils.py#165-166

Since we have this addition below, we don't have to do it again inside of this
if branch.

Differential Revision: https://phabricator.services.mozilla.com/D192067

UltraBlame original commit: 5e0414b6e7a9fa6193d4558f3715934aab7c7891