Bug 1845364 - Enable a11y in MacOS when role is requested from application. r=morgan
This is needed for Voice Control. Apple recommends it, and Chromium does this. In the future we may choose to have a more "basic" a11y mode toggled here that isn't as extensive as what we have for VoiceOver.

I couldn't figure out a way to detect Voice Control so we could include it in the telemetry/crash reporting.

Differential Revision: https://phabricator.services.mozilla.com/D184523

UltraBlame original commit: 7e8fd83cad4cec68465e006e23a7076f55275817