Bug 1730119 - Rewrite XSLT path result compare to use Collator::CompareStrings r=dom-core,hsivonen
Use `Collator::CompareStrings` instead of sort keys (see bug for more info).

For now I just removed `mCaseKey`/etc as it is unclear if it has a use
today, if so I can add it back in when we have the info on it.

Differential Revision: https://phabricator.services.mozilla.com/D192349

UltraBlame original commit: a5931e5d96c194965c007103e0692ecabaf6747c