Bug 1826504 - [devtools] Remove react-transition-group/Transition from Debugger vendors.js. r=devtools-reviewers,ochameau.
The library was only used to make the quick search panel
and the shortcut modal to show with a slide-in animation.
This can be replaced with a simple CSS animation instead.
I'm taking this as an opportunity to cleanup a few things
that were unecessary or not ideal in the CSS file.

Differential Revision: https://phabricator.services.mozilla.com/D192266

UltraBlame original commit: 839d107030fb1827d316e5f3c349f008611458c2