Bug 1858076 - [devtools] Show notification in the console when original variable mapping is disabled r=devtools-reviewers,nchevobbe
This patch displays a notification in the console when the debugger is paused
in an original file with original variable mapping is turned off.

The notification is displayed when console input is focused, also if the user selects
a generated file, or is no longer paused or switches on varaible mapping, the notification is removed
once the console input is refocused.

Also added support for specific notifications to control displaying the close button.

Differential Revision: https://phabricator.services.mozilla.com/D191568

UltraBlame original commit: 492052d366fe9bf860b43a33c528413f6692647f