Bug 1859055 - Make ShoppingProduct methods without a product argument static. r=calu,shopping-reviewers,jhirsch
Updates the `request`, `ohttpRequest`, `getOHTTPConfig`, `requestImageBlob` and `sendAttributionEvent` methods to be static, as they don't rely on having a product set on the class instance or passed as an argument.

Also adds passing an `AbortSignal` as an option for `request` and `requestImageBlob` instead of using the instances `AbortController`.

Differential Revision: https://phabricator.services.mozilla.com/D192411

UltraBlame original commit: 45d2c2d6343e237f1890eba5bd19d5b37ba133e6