Bug 1861566 - Allow some missing generated source files. r=firefox-build-system-reviewers,sergesanspaille
When gtest is not linked as part of the build, the source files that are
generated for it are not necessarily created during the build, so we
won't be able to find them when packaging them. We make this an allowed
situation. Ideally, we'd "just" figure out which files these are, and
not emit them in the list of generated sources, but at the moment, it's
a lot of work to be able to do so.

Differential Revision: https://phabricator.services.mozilla.com/D193003

UltraBlame original commit: 2bee4ca6cc38f674fa8d9848019482ade305fba7