Bug 1861442 - Avoid updating fields to 'cc-name-*' if a security code field is already present in the reference credit card section r=jneuberger
It is quite common for the security code field to be located at the end of a credit card section.
This implies that if we have already encountered a security code field, the subsequent fields are unlikely to
be part of the credit card section.

Differential Revision: https://phabricator.services.mozilla.com/D192562

UltraBlame original commit: 5218fc7cf52bb0d59a5c891cd7ccad4a2677a5ab