Bug 1863675 - Change how mozharness handles output_timeout. r=releng-reviewers,gbrown
The current code hits a deadlock when the timer thread kills the
subprocess, while the main thread is blocked on reading a line off the
subprocess's output: the readline call never returns, and while
mozharness prints the timeout message, it also stays there blocked, and
workers then wait for the task timeout, which is usually much larger.

The new strategy is to read from a thread, and send the read lines to
the main thread over a Queue, with a timeout.

Differential Revision: https://phabricator.services.mozilla.com/D193137

UltraBlame original commit: 543c915fc8c6775cd665ad6ebe58df0f46e9198c