Bug 1856982 - Add a way to add extra threads on top of default threads list to profile r=perftest-reviewers,kshampur
Previously we had the `gecko_profile_threads` argument to pass the threads to
be profiled. It's a useful argument but whenever users have to add an extra
thread, they have to add all the default threads to the list as well, like
"GeckoMain", "Renderer", and "Compositor".

This new argument allows users to add extra threads without having to specify
the default ones all the time. It appends the threads that are defined here to
the threads that are defined in `gecko_profile_threads`.

This way, for example, we can add "DOM Worker" thread to a specific test,
withou having to write down all the other threads.

Differential Revision: https://phabricator.services.mozilla.com/D192466

UltraBlame original commit: 2c0620057b0f3347b0d0571a5923302974f46424