Bug 1856301: [OpenBSD] also preload libavcodec.so before sandboxing r=gerard-majax
With the current sandboxing configuration, FFmpegRuntimeLinker::Init()
can't access libavcodec.so, so video decoding fails in the RDD process
and fallbacks to the content process.

With media.allow-audio-non-utility knob (from bug #1850904) toggled
to false enforcing audio decoding in the Utility process, this has the
side effect of enforcing video decoding in the RDD process, and thus
video decoding fails.

Preloading libavcodec.so like it was done for libmozavcodec.so and
libmozavutil.so in bug #1713745 fixes it, as the second dlopen() call
hits the ld.so cache and returns the preloaded library handle.

replaces/supersedes https://phabricator.services.mozilla.com/D190475

Differential Revision: https://phabricator.services.mozilla.com/D193353

UltraBlame original commit: 839befe903d4f2831853f5ce1f07f4dc74fa823d