Bug 1858858 - Handle data-l10n-id updates for <panelview> title. r=anti-tracking-reviewers,bvandersloot
The `<panelview>` data-l10n-id is often set immediately before rendering the subview. If the subview has previously been rendered, its title may be obsolete, and so the value needs to be checked directly -- the later l10n DOM update won't be caught by the panelview, and in any case we want the header to be set correctly before we show it.

Differential Revision: https://phabricator.services.mozilla.com/D190920

UltraBlame original commit: 7749507237fe4f34e6bc5711cedec1f3b14f335d