Bug 1863006 - [devtools] Tweak css classes for rules view selectors. r=devtools-reviewers,ochameau.
`ruleview-selectorcontainer` is renamed `ruleview-selectors-container`,
as it can holds multiple selectors.

`ruleview-selector` is renamed `ruleview-selector-element` as it is a part of
a selector, and this better match the classes we add for other parts (e.g. ruleview-selector-attribute).

Finally, we add a `ruleview-selector` class for the element that holds all the
parts of one selector, and we add `matched`/`unmatched` on it.
This class will also be used in the ancestor data (next patch in queue).

Tests that were using those classes are updated.

Differential Revision: https://phabricator.services.mozilla.com/D192864

UltraBlame original commit: cc98109f18578224560889d3998436d856458f49