Bug 1844958 - [devtools] Strike through unmatching selectors and var() args in Rules view. r=devtools-reviewers,ochameau.
We used to only differentiate those items with a dimmer color,
but it's not accessible to convey state change only with color (or it
needs to be way different).
So let's strike through those to indicate they're not used, which is
consistent with what we do for overidden properties.

Differential Revision: https://phabricator.services.mozilla.com/D192553

UltraBlame original commit: e8726c1480b933b358e8e6918a19c585640f1cfc