Bug 1862313 [wpt PR 42871] - Ensure restarted animations don't clobber start time, a=testonly
Automatic update from web-platform-tests
Ensure restarted animations don't clobber start time

The bug here occurs because of a race. Blink Animations can be
invalidated for all sorts of reasons and they require cancelling the
current cc::Animation and starting a new one. When this happens, the new
cc::Animation will use the start time that was received for the original
Animation.

The race condition occurs when the Blink Animation is invalidated before
a start time is received. In that case, the Animation starts playing on
the impl thread but the new Animation is created without a start time,
leading to the Animation starting from the beginning, causing a stutter
effect. Additionally, animation events generated on the impl thread when
Blink invalidates an animation are dropped since the main thread
counterpart is gone by the time of the events arrive.

This CL fixes the issue by making Blink recreate the compositor
animation with the same id and setting a bit on it to make it reuse
the start time from the compositor animation when it is pushed. The
animation also reuses the compositor group id on a restarted animation
so that animation events are routed back to the blink::Animation
correctly.

Bug: 1445090
Change-Id: I8e91bdbf8e22b00cd6025d21aa2a982ad4742d31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4605129
Commit-Queue: David Bokan <bokanchromium.org>
Reviewed-by: Robert Flack <flackrchromium.org>
Cr-Commit-Position: refs/heads/main{#1218221}

--

wpt-commits: 0fef73f2528f28ed7e8ea0c203f38ca092027f85
wpt-pr: 42871

UltraBlame original commit: 6cd48c6a5b6d8f824e710990adf544573d13f0a6