Bug 1843851 - Add a tp6 benchmark test to raptor. r=perftest-reviewers,kshampur
This patch adds a new "TP6 Benchmark" test that quickly runs through all of the test pages we have in TP6. No visual metrics are recorded, and we attempt to gather the information into a single metric within the `browsertime_tp6_bench.py` support file. This test is in-progress and shouldn't run in production, but should be available on try, and locally so that we can continue developing it.

Depends on D183751

Differential Revision: https://phabricator.services.mozilla.com/D183752

UltraBlame original commit: 9517987e19427c2900046f62ca816d243d1f1403