Bug 1863995 - Make webtransport datagram test more robust, r=necko-reviewers,valentin
With neqo 0.6.7, we send ACK packet if the last time we ACK is more than one RTT. This change makes our test HTTP/3 server send an ACK packet before sending the webtransport datagram.
On windows platform, this causes a problem that the server failed to send the datagram, because ACK frame and datagram frame are in two UDP packets and the OS API prevents from sending two patckets in a short time.
Since webtransport datgram is unreiable, quic won't retransmit it. As the result, the test fails because of timeout.
To fix this problem, this patch uses another way to check if the server really gets the datagram sent from Firefox by creating another HTTP request to get the stored datagram.

Differential Revision: https://phabricator.services.mozilla.com/D194046

UltraBlame original commit: 7626d9d131fb4daff7deedc79b7aa27eef41aa25