Bug 1864554 - Reimplement ScrollAnimationMSDPhysics::ApplyContentShift() to reflect the content shift in the physics models. r=hiro
In bug 1759958, we tried an implementation that tacks on the content shift
at the end, but this is not consistent with Update() being called with
a destination that takes into account the content shift.

This fixes APZCSmoothScrollTester.ContentShiftDoesNotCauseOvershootMsd.

Depends on D193978

Differential Revision: https://phabricator.services.mozilla.com/D193979

UltraBlame original commit: 6b5c80f3fd764927e6c8dadaeea14ab6209e426f