Bug 1861916 - Don't fail for Protocol attribute errors on Python < 3.10, r=Sasha
We previously tried to skip tracebacks for these failures by checking the obj property
on the error object. But that turns out to only work for recent Python versions, so
on older versions fall back to inspecting the error string to figure out which
object raised the exception. This is a bit of a hack, but it's probably OK to
solve the problem since it only really affects logging.

Differential Revision: https://phabricator.services.mozilla.com/D192829

UltraBlame original commit: d238dc5f29bd0251003e86f580984cf1171d8729