Bug 1813128 - Part 1: Populate the rule id to the clicking rule. r=pbz
To collect the telemetry regarding CMPs, we need to know which CMP rule
is used for handling the cookie banner. So, we populate the rule id to
the click rule, the rule id contains the name of the CMP for global
rules. We will use the id to report telemetry.

Differential Revision: https://phabricator.services.mozilla.com/D193705

UltraBlame original commit: 6f63c50d10b5b8c233a8e226f48089f480e881d6