Bug 1839304 - Don't disconnect Places result from the treeview on setTree(null). r=daisuke,places-reviewers
SetTree(null) may be invoked in various cases where the view is temporarily
hidden, but not unloaded. the Places view should stop listening to changes only
when the document goes away, not when the tree is unset. In that case it should
be sufficient to close the root container to free up memory and avoid handling
notifications.

Differential Revision: https://phabricator.services.mozilla.com/D194640

UltraBlame original commit: 66c2fb78dfe25ce9af21d0e1dd82f62dbb5da53f