Bug 1861734 - Disable v8 context snapshot in CaR builds. r=perftest-reviewers,afinder
Sometimes a CaR build will fail due to being unable to use a v8 context
snapshot. It is unclear why or when this intermittent issue happens and
is hard to replicate. Previously it occured on windows consistently and
the `use_v8_context_snapshot` build flag was set to false.

To be safe this patch sets this flag for all platforms to hopefully
prevent this from happening again.

Differential Revision: https://phabricator.services.mozilla.com/D194652

UltraBlame original commit: 4c5b2b4298575e99e99d7203ecdb33fe66c65358