Bug 1858225 - [4/9] use alternate-thread file picker in remote process  r=win-reviewers,handyman,mhowell
When opening a file picker in a utility process, do so off the utility
process's main thread, allowing multiple dialogs to resolve in parallel.

The relevant IPC calls are already formally async, so this suffices to
convert them to being functionally async as well. (While this is
technically a functional change, it's not yet a user-visible one: the
parent-process code is still synchronous.)

Differential Revision: https://phabricator.services.mozilla.com/D193736

UltraBlame original commit: 9b2c6eaf137e0050ee395a704500b0721c0fe3b6