Bug 1867170: a11y checks: Check the correct node in some cases involving shadow DOM. r=ayeddi
First, if a click is synthesised at the centre of a custom element containing a slot, the click might target a text node which is a child of the slot.
In that case, the composedTarget for the click event will be the slot, since click events on text nodes are retargeted to the parent node.
The slot itself isn't rendered, so we were returning early due to the 0 bounds check, hiding potential failures.
To fix this, walk to the slot's flattened tree parent.

Second, if a click occurs on a node which doesn't have an Accessible, we try to walk to the nearest Accessible ancestor.
If the node is in shadow DOM, the nearest Accessible might be outside of the shadow DOM.
Previously, we would stop at the shadow root.
Now, we walk flattened tree parents, which will walk us out of the shadow root.

Differential Revision: https://phabricator.services.mozilla.com/D194944

UltraBlame original commit: ce5a379c35bead2efe1e4327e6230b8ba278b1c9