Bug 1858225 - [5/9] use local-async functions for local calls  r=win-reviewers,mhowell
When opening a file dialog locally (that is, in the parent process), do
so asynchronously on a separate thread.

Since the calling code is still synchronous, this necessitates adding
new calls to `ImmorallyDrivePromiseToCompletion` at the sync-to-async
boundary. These are _very_ temporary -- they'll be removed before this
patchset is done.

Differential Revision: https://phabricator.services.mozilla.com/D193737

UltraBlame original commit: d08b97ae0098e4be4ecd4415eccdf996a0d7cdb0