Bug 1858225 - [6/9] hoist syncification one layer up  r=win-reviewers,handyman
Pull the sync-to-async boundary up one function-call layer, from
`ShowXxxPicker{Local,Remote}` to `ShowXxxPicker`. This coalesces the
remote and local sync-to-async boundaries.

Additionally, standardize on exclusive MozPromises, since none of them
happen to need multiple listeners.

This is arguably, but not practically, a functional change, as a tiny
bit of glue code is now executed on a different thread.

(N.B.: While this reduces the total number of sync-to-async boundaries
back down to two, this is not the removal promised in the previous
commit summary.)

Differential Revision: https://phabricator.services.mozilla.com/D193738

UltraBlame original commit: e66d99182863ee02a5e4455be7cc6f3a11578e6e