Bug 1867673 - Gracefully handle statx when called in sandboxed GMP processes. r=gcp
Testing note: this bug seems to break all media plugins on 32-bit x86
with a sufficiently recent glibc, including the clearkey plugin, so
the EME tests we already have provide coverage.  However, we no longer
test 32-bit linux on CI, and we may not yet have a new enough glibc
there either.  As for testing the syscall directly, we have a test suite
for that, but it runs with crash-on-error mode disabled and this patch
doesn't change the observed behavior in that case.  Thus, there is no
test case in this patch.

Differential Revision: https://phabricator.services.mozilla.com/D195194

UltraBlame original commit: 97bd03fdb1bec86b388b3393aa328608da4b388d