Bug 1858279 - Prelude: change Unix LaunchApp to take LaunchOptions by move. r=nika
The next patch will need the Linux version of LaunchApp to side-effect
the LaunchOptions in some cases; given that none of the callers ever
try to reuse a LaunchOptions, we might as well change it from a const
reference to an rvalue reference.  The Mac version is also changed,
because it uses the same declaration / function type.

Also, the overload of LaunchApp that takes a CommandLine has been
removed, because it was unused.

However, the Windows version is *not* changed; there's no need for it,
and it accepts different types of arguments anyway so it's not necessary
to match the Unix one.  Also, the CommandLine overload of the Windows
LaunchApp *is* used and continues to exist.

Differential Revision: https://phabricator.services.mozilla.com/D194455

UltraBlame original commit: b630720092110140777657cd687f0073c1d819dc