Bug 1609403 - Grid items with align/justify-self should participate in baseline alginment r=emilio,layout-reviewers
When a grid item is set to align itself to the baseline (first or last)
then it should participate in baseline alignment if the item's size is
not dependent on the size of the track and vice versa.

Also added a test to check for last baseline alignment in addition to
first baseline alignment on empty blocks.

Differential Revision: https://phabricator.services.mozilla.com/D157345

UltraBlame original commit: 02d58cff5d1ab112ebbe2ad15cb58f9263684fb9