Bug 1865421 - Move from static inline to proper functions to get a lock. r=chutten
Due to the functions ending up in a header and marked as `static inline`
they were potentially inlined.
Which in turn leads to the locked map actually not being a shared locked
map, but one per invocation.
Which in turn meant that every time the Counter implementation tried to
look up the scalar ID for a submetric it found nothing and thus recorded
nothing.
The new test catches that (it fails as expected on the old code and
passes on this commit).

This also has the benefit that changing this functions does not require
to rebuild the world anymore!

Differential Revision: https://phabricator.services.mozilla.com/D194371

UltraBlame original commit: 5226b98a51ca9d5a19f379645efa4fa04678a738