Bug 1859549 - Remove from condprof util.py call to install unpacked xpi in the generated condprofile. r=sparky
While investigating a frequent timeout failure that the new tp6m webextensions task was hitting in the preliminary push to try,
the try push were already showing that it was hit more often when using the settled-webext conditioned profile then how
often was hit so far with the settled conditioned profile, I have recalled that in condprof util.py we are installing the
addon as an xpi file and also as an unpacked dir, that turned out to be the reason why running the task was taking often
more than 3 minutes with no output (due to browsertime being running quite a number of adb commands for each of the files
in the unpacked extension directory).

On the AddonManager side we expect to be installing the addons sideloaded in the profile as packed xpi files,
and in general unpacked extensions to be loaded temporarily, and so given that it doesn't seem any of the condprofile
customization file is using the "addons" property (besides the new webext.json one we are introducing in this
stack on patches), this patch just removed that additional call that is unpacking the extension xpi file in the
condprof profile.

If we ever need that for a reasonable use case, we may consider adding it back as an opt-in feature (e.g. through a new
property to be explicitly added in the customization json file).

Depends on D194637

Differential Revision: https://phabricator.services.mozilla.com/D195664

UltraBlame original commit: 4fe34077c984323ae55175f3cbf7b7d95d4c7608