Bug 1868382 - Check if a provider belongs to an allow-list if it isn't exported when reading pdfs in GV r=geckoview-reviewers,owlish
Some Google apps which are files provider are exported or not, hence checking for
the exported flag is likely too restrictive.
So for the case where exported flag is false, we check that the authority starts
with a string belonging to an allow list.

Differential Revision: https://phabricator.services.mozilla.com/D195556

UltraBlame original commit: 638dd51545aa51bea181e2aec835f52636d9594c