Bug 1863914: Use multiple shmem buffers for remote canvas recording. r=aosmond
This replaces the use of a single large ring buffer.
The buffers are still processed in parallel and are recycled to reduce
allocation. Events that do not fit in the default sized buffer have a separate
buffer created to fit them. These large buffers are not recycled.
Separate shared memory is used for readback, with a single shmem cached for this
purpose. Generally only one cached shmem should be required, because the
operations that usually readback the data do it straight away.

Differential Revision: https://phabricator.services.mozilla.com/D193207

UltraBlame original commit: 8839a3ba62436bd0ea7e5bd07e2b63a8d45b8609