Bug 1867193 - Fix baseline-compiled check in MaybeCreateAllocSite. r=iain
Allocation sites for IC stubs are created either when we Baseline compile the script,
or when we attach a new stub after Baseline compilation.

When attaching a stub we were checking whether we have an interpreter frame, but that
fails when we have a Baseline script but are currently running in the interpreter.
This can happen after a bailout for example. In this case we'd never create the allocation
site.

Differential Revision: https://phabricator.services.mozilla.com/D195895

UltraBlame original commit: c148a4f270cac68e05cbb78152305e9279c16f6a