Bug 1867193 - Move active flag to ICScript, and add bytecode size field to ICScript. r=iain
The active flag is currently set on the JitScript, but the next patch will also
use this to mark trial-inlined ICScripts that are on the stack. We can use this to
preserve those trial-inlined scripts and discard the rest.

To properly discard the trial-inlined ICScripts we also need to know the bytecode size
of the script, to update `InliningRoot::totalBytecodeSize_`, so this patch also adds
this to the IC script.

Differential Revision: https://phabricator.services.mozilla.com/D196172

UltraBlame original commit: 0e24a59eae533518f0aab0905157258eaa18c642