Bug 1867296 [wpt PR 43417] - [css-images-4] Implement unprefixed cross-fade()., a=testonly
Automatic update from web-platform-tests
[css-images-4] Implement unprefixed cross-fade().

This is similar to -webkit-cross-fade(), but differs in that it
has a different syntax, and supports more (or less) than two images.
It also supports cross-fading against a solid color, which we implement
internally as a gradient for simplicity. (This may not be optimal for
memory or performance, but it should do for now.)

We do not implement parts of the spec that are not yet written
but will be, in particular simplification at computed value time.

The functionality is behind a new runtime flag CSSCrossFade,
which is set to experimental as usual.

Change-Id: I4d29e14b215a65223b98dcaee8ba99b8f62d94dc
Bug: 614906
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5055552
Commit-Queue: Anders Hartvoll Ruud <andruudchromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruudchromium.org>
Auto-Submit: Steinar H Gunderson <sessechromium.org>
Cr-Commit-Position: refs/heads/main{#1233870}

--

wpt-commits: 2f00e95d38dbb8969039840432c5fb42f3c23255
wpt-pr: 43417

UltraBlame original commit: 20e3f7c6c888373e780dbe2091b36005c548189c