Bug 1869569 - Avoid CompositorOGL::BeginFrame's glInvalidateFramebuffer optimization on Xclipse 920 GPU. r=gfx-reviewers,nical
In CompositorOGL (used by software webrender on Android) as an
optimization we use glInvalidateFramebuffer rather than glClear for
the color attachment if there is no non-opaque invalid region. This
causes rendering glitches on the Samsung Xclipse 920 GPU. This patch
marks GLfeature::invalidate_framebuffer as unsupported on this GPU,
causing us to fall back to the glClear pach which avoids the glitches.

Differential Revision: https://phabricator.services.mozilla.com/D196416

UltraBlame original commit: 9094e6aea726211862216b3538cd1b309580105c