Bug 1869558 - Port applicationServerKey validation tests to WPT r=asuth
Porting only the main thread checks but not the checks inside service workers, because it's not clear what to do with permissions there as testdriver.js is window specific. Maybe it can be tweaked for service worker compatibility, but not in this patch.

Differential Revision: https://phabricator.services.mozilla.com/D196181

UltraBlame original commit: f9c41a2aac8b768d747b876d4363314e39cd07c3