Bug 1870417 - Use 64-bit arithmetic to avoid overflow during multiplication. r=gfx-reviewers,lsalzman
This avoids the multiply-with-overflow crash here, though the dimensions in the testcase
are so vast that it then hangs trying to process thousands of tiles at once (as seen in
bug 1864374), even after increasing the tile size.

For cases like this, we really need to clamp maximum sizes somewhere earlier in the
rendering pipeline. But that's what bug 1864374 is already about, so here let's just fix
the overflow issue.

Differential Revision: https://phabricator.services.mozilla.com/D196688

UltraBlame original commit: ca7546c202a66b0b8985da70b872418c70312eab