Bug 1870458 - Make dom/bindings/test/test_sequence_wrapping.html not interfere with mochitest framework. r=dom-core,edgar
This test adds a numeric property to Array.prototype, which then causes issues
in the mochitest framework when it tries to add things to an empty array while
gathering the results. Doing the test's work in an iframe avoids the issue.

Differential Revision: https://phabricator.services.mozilla.com/D196662

UltraBlame original commit: d614e05cf543d76028dff08af829fafa5f45cfaf