Bug 1848829: Do not check pref sanitization for Static Prefs r=nika
If `IsPreferenceSanitized(char*)` is called, it will call pref_Lookup
which is not thread-safe.  StaticPrefs use that function to check
pref sanitization.

Prefs are sanitized if they are a no-default-value string-type preference
_or_ their name is in sRestrictFromWebContentProcesses. The first check
is intended to catch prefs set dynamically, ones we don't know about
and may want to sanitize. But Static Prefs we know all about. So we can
omit that check for them. We only (in theory) need to check them against
the blocklist.

In practice: no Static Prefs are sanitized.  If they were (and they were
not mirror-never prefs), it would have exposed the bug that makes content
processes crash on startup, because we would crash during initialization.

That initialization happens while reading the prefs from the Shared Map -
sanitized prefs can't be in the shared map. So we can take a shortcut and
instead of building in support for Static Prefs to be initialized from
the HashMap instead of the shared map (perhaps as simple as
s/GetSharedPrefValue/GetPrefValue), we can just crash if someone adds a
Static Pref to the list of things to be sanitized.

Differential Revision: https://phabricator.services.mozilla.com/D189968

UltraBlame original commit: 7e186e1a46085a53c7e5b80e45e88cee3daecb65