Bug 1870519 - Add level to H264Specific r=media-playback-reviewers,alwu
This patch returns the H264 level with the profile.

Note that if profile and level cannot be parsed from the parameters, the
default (profile, level) is (main, 31), which is different from the
(base, 31) suggested in `ParseSdpForH264ProfileLevelId`.

For now, it's probably ok, since level value is unused in WebRTC's
platform encoder for now.

Differential Revision: https://phabricator.services.mozilla.com/D196681

UltraBlame original commit: 71e23dd081167c51722e8b8d4c62ec40bea64020