Bug 1829026 - Don't force sync for async-present if GL not thread-safe. r=aosmond a=graft
If GL is not thread-safe, then the check for whether we use the canvas render thread
is not accurate. The thread-safe check takes precedence, and if we are not thread-safe,
we are actually on the render thread. It should be safe to not force sync here since
we aren't actually on the compositor thread at all.

Differential Revision: https://phabricator.services.mozilla.com/D196864

UltraBlame original commit: 887e5358bad02aa565b0a491ac97644f6bf32f53