Bug 1869878 [wpt PR 43651] - WebKit export of https://bugs.webkit.org/show_bug.cgi?id=260201, a=testonly
Automatic update from web-platform-tests
WebKit export of https://bugs.webkit.org/show_bug.cgi?id=260201 (#43651)

Size containment is supposed to size a box as if it had no content, so
for the purposes of grid layout this would mean that we would want to
skip parts of the algorithm that take into consideration the content.

The "Stretch auto Tracks," portion of the grid track sizing algorithm
should not be skipped, however, as it has nothing to do with the content
of the grid. We should be able to distribute any definite free space
that we computed up to this point.
--

wpt-commits: ccb431fd5075cf19bee842c3a8c4661f38d2a066
wpt-pr: 43651

UltraBlame original commit: 425a2c0cad8eacdbeb8dd744be7e2037165b9983