Bug 1870652 - Remove unnecessary additionalProperties requirement in search-config-v2-schema. r=mcheang
Drop the requirement for disallowing additional properties in the order section as remote settings as its own properties, and this wasn't meant to be included.

We do check additional properties in xpcshell-tests, but need to allow for additional properties at runtime in case configurations are extended.

Depends on D196737

Differential Revision: https://phabricator.services.mozilla.com/D196738

UltraBlame original commit: 85d549eafb178fcd145a12e8d6b15994267e5687