Bug 1869242 - Part 2: Use unknown allocation site data to turn on per-zone string pretenuring r=sfink
The 'unknown' allocation site is used to track nursery allocations when we
don't have a sepcific allocation site. This is per-trace kind and per-zone, and
the data it collects can be used to make the decision on whether to disable
nursery allocation for a trace kind and zone.

This is better than our current system because it's specific to the trace kind,
and it also means we don't also have to collect data for the original scheme.
It also means we can do the same thing for big ints.

Differential Revision: https://phabricator.services.mozilla.com/D196273

UltraBlame original commit: 5ec713ef6a20a7103586e1d277bc30de16cf4103