Bug 1871188 - Ignore all warnings in mio-0.6.23. r=emilio
For all intent and purposes, it's a third party crate, and we don't deny
warnings in third party crates. It happens with mio 0.6 because we're
patching it to use winapi 0.3 and miow 0.3.

We also revert some older changes, to reduce the number of changes
against the canonical 0.6.23.

Differential Revision: https://phabricator.services.mozilla.com/D197026

UltraBlame original commit: 56f99da689892a9b800fda5fee225066c0c9c619