Bug 779684 - part 3: Make `nsFrameIterator` non-refcountable r=emilio
Nobody shares an instance of `nsFrameIterator`.  Therefore, we can make it
non-refcountable.  Additionally, `nsVisualIterator` is not required because
it overrides only 4 methods which are really simple.  So, once we merge it
into `nsFrameIterator`, we can allocate it in the stack.

Differential Revision: https://phabricator.services.mozilla.com/D197144

UltraBlame original commit: 8b92fb4e3f41d856629818bd2ecfc25bed23a643