Bug 1872087 - Share recycling across WebGL framebuffers on a single context. r=aosmond,gfx-reviewers
This was a change for bug 1872087 that somehow got lost when I was rebasing. The idea
here is that DrawTargetWebgls use a single shared WebGL context, so it makes sense for
them to share recycle texture data, much like CanvasTranslator shares recycling of
texture data for D2D. This should increase the efficiency of use-cases that create and
destroy a lot of canvases every frame and would not normally benefit from recycling.

Depends on D197375

Differential Revision: https://phabricator.services.mozilla.com/D197376

UltraBlame original commit: 2e15b5318d54850660d5c4e4ece3c82b2634d806