Bug 1870401 - Add duration tests r=media-playback-reviewers,alwu
These tests ensure the durations of the produced encoded data is greater
than zero.

The ffmpeg video encoder used to produce zero duration due to rounding
issues. The recommended time unit is *1/framerate* second, whie the time
unit in the tests (and in WebCodecs) are *microseond*. In the current
mechanism (ToTicksAtRate [1]), converting *X* microsecond, where *X* is
positive, into *Y* in *1/framerate* second doesn't guarantee *Y* is
positive as well. *Y* can be zero sometimes.

To work aroud this issue, ffmpeg video encoder currently set the time
unit to microsecond as well so there is no need to do the conversion.
These added tests are used to make sure no regressions will be
introduced when we are ready to take care of the rounding calculations.

Below are the details about the rounding problem:
-----------------------------------------------------------------------
In the test, the framerate is 30 frames/second, which means each
encoding iteration is 1/30 second. If we encode one frame per iteration,
the duration should be 1/30 second as well, which is roughly 33333
microseonds set in the test.

If ffmpeg video encoder set the time unit to *1/framerate* second, as
ffmpeg suggested, then we would need to convert 33333 microsecond into
*X* in *1/framerate* second. Since one microseond is (framerate/1000000)
in *1/framerate* second, *X* is (framerate * 33333)/1000000 =
999990/1000000. This calculation is actually what
`TimeUnit::ToTicksAtRate` does, when we need to convert the time from
one unit into another.

However, in praticem the *X* would be `0` because `ToTicksAtRate` always
round down the calculated results. That is, a 33333 microsecond duration
become zero duration at the end of the conversion.

[1] https://searchfox.org/mozilla-central/rev/593c49fa812ceb4be45fcea7c9e90d15f59edb70/dom/media/TimeUnits.cpp#92

Differential Revision: https://phabricator.services.mozilla.com/D197356

UltraBlame original commit: 697b731b550cc99bf7c7ddf060e36d4470e59299