Bug 1871596 - Add a programmatic role to the search-go-button image within the Searchbar. r=search-reviewers,Standard8
`search-go-button urlbar-icon` is an `<image>` element that has an accessible name but lacks an appropriate interactive role, because it functions as a button (not focusable with keyboard because the alternative - `Enter` key - exists and works for the search submission and we do not want to create an additional tab stop for keyboard-only users). This would prevent users of speech-to-text/Voice Control from being able to send a click to it by calling its label and screen reader users and users of other assistive technology would not be able to get to this control via shortcuts like a list of controls, etc.

This issue is similar to the `Go` button on the URL bar resolved in the bug 1864962

To activate the searchbar next to the URL Bar that includes this go button: `about:preferences` > `Search` > `Search Bar` > select `Add search bar in toolbar`.

Differential Revision: https://phabricator.services.mozilla.com/D197202

UltraBlame original commit: 8146df27da1a94bc1dacfae424bdb114ce7db97c