Bug 1655101 - AssertTiedFieldsAreExhaustive<T>() gives better compiler errors. r=gfx-reviewers,aosmond
A templated class FieldDebugInfoT is used, because compilers will print
the template parameters of classes involved in error messages, so we can
use this to print a number of pieces of info, which will pinpoint where
the issue is.

static_assert with constexpr formatting is coming in C++26. :')

Differential Revision: https://phabricator.services.mozilla.com/D197542

UltraBlame original commit: 260f3f01f197de1cfb3db72acef17f9c055490f2