Bug 1873147 - [devtools] Update token highlight and popup position when Popup updates. r=devtools-reviewers,ochameau.
In Popup, we were only adding/removing the highlight on hovered token on mount/unmount.
But it can happen that the component gets updated (e.g. when the user moves quickly
between valid tokens).
In such case, we want to remove the highlighted class on all elements that it was
set on, and only add it back on the "preview target".
We also want to update the Popover position so it points to the actual token the
preview was computed for.

Differential Revision: https://phabricator.services.mozilla.com/D197582

UltraBlame original commit: 5433f5de59c8ce8071874a9238e27f9bfb780a3e