Bug 1871024 - Update in-content border color to use border-color token. r=dao,desktop-theme-reviewers
This addresses the lack of visual separation in about:logins on Ubuntu
in High Contrast mode. The about:logins sidebar having the same
background color as the main body of the page is expected in High
Contrast mode, so we re-establish the borders on the sidebar to provide
visual separation between the elements.

Differential Revision: https://phabricator.services.mozilla.com/D197728

UltraBlame original commit: 88cd4ff7de1e48ee772d73b266239dd8b7adab1f