Bug 1822630, step 2, part 1: test(webgpu): set effectively inf. depth for CTS test grouping r=webgpu-reviewers,jimb,jmaher
WebGPU CTS testing in `mozilla-central` is a set of WPT tests that are
generated from WebGPU upstream. Because we intend to generate test files
with relatively fast execution times, we on the WebGPU team believe that
it no longer makes sense to adhere to the current chunk division logic
in the decision task. We are currently unaware of a better way to do
this, but believe this change to be uninvasive enough to propose as
a patch first.

Differential Revision: https://phabricator.services.mozilla.com/D196627

UltraBlame original commit: 009ad7a2177f25e995cf2e1affa069f76b9caa7c