Bug 1859608 - Change the user's DoH region when that changes r=necko-reviewers,sunil,kershaw
The DoHConfig now listens for region change notification and
saves the new region in a pref. On the next initialization of
the DoH config we will use the saved pref to update the doh
region even if the pref was already set.

This should also work for testing (manually setting the value of the
pref) as long as no region changes happen.

Differential Revision: https://phabricator.services.mozilla.com/D195023

UltraBlame original commit: fea9dbbad1ab1a14313510023f859abb578c7fbf