Bug 1870157 - Provide programmatic highlighting to Highlighted fields on the Address capture doorhanger r=dimi
When the autofilled address is edited, the Address capture doorhanger appears with the fields that were changed being highlighted. This is not communicated to an assistive technology because it is marked up as `<span>` that has no programmatic meaning/role.

Differential Revision: https://phabricator.services.mozilla.com/D196967

UltraBlame original commit: 6df5e64d0de12c44cb3a0036e3ed43f24ac2a1ef