Bug 1872986 [wpt PR 43849] - [scope] Produce kScopeActivation relations for scoped nesting, a=testonly
Automatic update from web-platform-tests
[scope] Produce kScopeActivation relations for scoped nesting

When parsing style rules that are immediate children of an scope rule,
any compound selector which contains either :scope or '&' is prepended
with RelationType::kScopeActivation. This relation is what triggers
the "figure out if we're indeed in scope, and what the scoping root is"-
behavior during SelectorChecker, and without the kScopeActivation, we'll
just fall back to the regular behavior for :scope (i.e. it matches :root
in the normal cases).

This CL fixes a bug which caused missing kScopeActivations for nested
style rules within scope, e.g.:

  scope (a) {
    b {
      /* :scope needs to refer to 'a' here */
      :scope:not(&) { ... }
    }
  }

This doesn't work, because we're basing the kScopeActivation emit/no-emit
decision on the nesting type of the *immediate* parent rule, which in this
case is CSSNestingType::kNesting. We need to know whether any ancestor
rule is an scope rule, and so we propagate a flag for this, and base
the emit/no-emit decision on this instead.

Fixed: 1512217
Change-Id: I35d28489dd8255badd2e9a9c911fda3e2541bb52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5164291
Reviewed-by: Rune Lillesveen <futharkchromium.org>
Auto-Submit: Anders Hartvoll Ruud <andruudchromium.org>
Commit-Queue: Rune Lillesveen <futharkchromium.org>
Cr-Commit-Position: refs/heads/main{#1242853}

--

wpt-commits: 1ebd9648f556cf120a941d0b2d44b6bfb23f67ae
wpt-pr: 43849

UltraBlame original commit: afd101f7f29ecd907c0b74855c720e3365d54e4b