Bug 1829457 - Cache DrawTargetWebgl snapshots for fallback when context lost after minimize. r=aosmond
After a minimize, an unknown amount of time or circumstances may be involved that ultimately lead to
a GL context loss. To try to mitigate this, cache software snapshots of DrawTargetWebgls when we are
about to minimize so that these can hopefully be copied into fallback TextureDatas later if the context
is actually lost.

Differential Revision: https://phabricator.services.mozilla.com/D198129

UltraBlame original commit: 112b960c1471cb1ea991442d85bad0cff91185f4