Bug 1869043 add a main thread record of track audio outputs r=padenot
mAudioOutputParams will also be used in a subsequent patch to support
secondary output devices in the MediaTrackGraph because a CrossGraphReceiver
will be created and destroyed on the main thread when it knows whether or not
a secondary device will be in use.

Volumes of different audio outputs (with different keys) on the same track are
coalesced on the main thread, so that the graph thread need not care about
output keys, nor mix the same stream multiple times for different keys.

Differential Revision: https://phabricator.services.mozilla.com/D198225

UltraBlame original commit: ae005b3084dcfa4f288d6f02d3560889abdeba8b