Bug 1873149 - [devtools] Don't show preview popup for keywords. r=devtools-reviewers,ochameau.
We shouldn't show the popup for keywords (`const`, `let`, `for`, …).
codeMirror marks those with the `cm-keyword` class, so we can ignore those in `_isInvalidTarget`,
but we need to exclude `this`, for which we do want to show the preview.

Differential Revision: https://phabricator.services.mozilla.com/D197585

UltraBlame original commit: 22d8df4c662674b7eb80d2b926a3bbe44da70472