Bug 1862022 - [devtools] Fix the scroll issue in the quickopen panel r=devtools-reviewers,nchevobbe
This fixes the issue allowing the result list to scroll as the user presses the up/down to move through
the list items. Seperating this patch from the test (as the test is quite tricky), so we can land the fix.

Differential Revision: https://phabricator.services.mozilla.com/D193721

UltraBlame original commit: 8456ea9fb109f770671bd8f507a4196b1ccbf2c6