Bug 1874254 - Allow decompression of both x86 and ARM BCJs on all platforms r=nalexander,mhowell
Without this change, MARs compressed with (for example) x86 BCJ filters fail to decompress on Apple Silicon. This change isn't strictly _necessary_ in that we are not planning to need this support, but it has the potential to prevent problems in testing and production. The consequences of turning it on should be minimal.

Differential Revision: https://phabricator.services.mozilla.com/D198443

UltraBlame original commit: d39befe31ff87f71d9c5b8a51d07e7dc867072f2