Bug 1870459 - Make legacycaller work through Xrays. r=jandem
HTMLAllCollection is marked as legacycaller (which means it supports doing
document.all(anindex)). But it also has an indexed getter, so we implement it
using a proxy. Proxies don't have a call op in their JSClassOps, so we need to
use BaseProxyHandler::call for them.

Differential Revision: https://phabricator.services.mozilla.com/D196663

UltraBlame original commit: 45d6750560827745ad0db7430dc5e8546ce28f6c