Bug 1875404 - Require branch for setPref variables r=chumphreys
The setPref variable annotation now requires both the pref to set and the
branch that should be used. Previously this was determined by the
isEarlyStartup feature annotation: true meant that prefs would be set on the
user branch and false meant that they would be set on the default branch. All
setPref annotations have been updated to stay consistent with that model.

Differential Revision: https://phabricator.services.mozilla.com/D199062

UltraBlame original commit: 1cc2adeae051b1fff55a31ece048718eadeaa66d