Bug 1870500 - Resolving the WPT intermittent failure. r=perftest-reviewers,kshampur
The issue we were encountering here was a missing metric for canva.com intermittently occuring
That ocured because of the location of that metric(first contentful paint) in the returned json moved locations for just canva.com, and so canva.com was removed whilst the problem debugged. Afterwards it was discovered that this issue was covering up another intermittent about json decoding error which was in turn resolved by removing code which removed "www." from links used with python requests.

Differential Revision: https://phabricator.services.mozilla.com/D199238

UltraBlame original commit: a0389da68b91e5a92b02ccd9e65991bffb5d5771