Bug 1875424 - Make `EventStateManager` dispatch `mouseleave` events on ancestors even after the last `mouseover` target is removed r=smaug
Currently, `EventStateManager` clears the last `mouseover` element when it's
removed from the tree.  Therefore, it does not fire `mouseleave` events when
the cursor moves outside of ancestors which a `mouseenter` event is fired on.

This patch makes `EventStateManager` keep storing the parent of removing content
node which is an inclusive ancestor of the last `mouseover` target.

The following patch will make the code easier to read.

Differential Revision: https://phabricator.services.mozilla.com/D199198

UltraBlame original commit: 997aa1f6b4cc89069e8b1e929772a06f4adefa1f