Bug 1875781 - Add explicit non-default URLSearchParams deserialization; r=dom-storage-reviewers,asuth
This is a preparation for bug 1636761 which needs to make URLSearchParams not
structured-cloneable. The deserialization uses the existing method
URLSearchParams::ReadStructuredClone for now, but it will be replaced with own
deserialization in bug 1636761 because URLSearchParams::ReadStructuredClone
will be removed.

Differential Revision: https://phabricator.services.mozilla.com/D199344

UltraBlame original commit: ad0e7eb94dfce3a7e15ec1530f253926d456d986