Bug 1876913 - part 1: Port `test_bug1406726.html` to WPT r=m_kato
This bug is caused by that we're still using the traditional behavior for
hack in `InsertParagraphSeparatorAsSubAction` to wrap current line with the
default paragraph separator here:
https://searchfox.org/mozilla-central/rev/07de26de179b62f9663e3cafff21874b55acf026/editor/libeditor/HTMLEditSubActionHandler.cpp#2021-2031

If we make it use the format block mode, it starts to refer the computed style
to consider whether block or inline for the blocked `<span>`, but that changes
other behaviors too and `test_bug1406726.html` starts failing and the
expectations are really different from results in Chrome.  For the further
work, this patch ports the tests into WPT and align the expectations to Chrome.

Differential Revision: https://phabricator.services.mozilla.com/D199844

UltraBlame original commit: 29a6fd68ebb0968e5e96b4f0b2f66b71ca789a7c