Bug 1877180 - Pass the app name to mozproxy during recording. r=perftest-reviewers,kshampur,afinder DONTBUILD
This patch fixes a couple bugs with recording. The first is that the app wasn't being passed to mozproxy, and this was preventing the creation of the mitmproxy certificate for the app being used to record. The second is that the --proxy-deterministic flag isn't defined within the proxy layer even though it's used by it from the hooks.

Differential Revision: https://phabricator.services.mozilla.com/D199891

UltraBlame original commit: 85d4a93d2fce8e976fce8c0e6ae573731793a21d