Bug 1749345 - Skip FirstStartup initting if a previous profile exists. r=nalexander,rhelmer,aminomancer
The thinking here being that if a previous profile exists, then the
--first-startup argument is probably be passed because the user is
reinstalling on a system that still has (or once had) the browser
already installed on it. In that case, we're going to use that
pre-existing profile, and we don't need to do the FirstStartup
things, since they're primarily for systems where a new profile
is being created after install.

Differential Revision: https://phabricator.services.mozilla.com/D199763

UltraBlame original commit: 3908f0d77cd1e925431fddd7449a89398947fce7