Bug 1875670 - Create a refcounted object for resolved color-font palette, instead of just a bare nsTArray. r=gfx-reviewers,lsalzman
This does not change behavior for users; it is preparation for the
following patch, which will want the ability to hold an additional
strong reference to a resolved palette. So we can no longer just own it
via a UniquePtr; we make it a RefPtr instead.

Differential Revision: https://phabricator.services.mozilla.com/D200138

UltraBlame original commit: 649109bd35ad159d4fea019aa39c1956a07f7bc0