Bug 1876421 - Update AccessibilityUtils to check for menuitems when calculating an accessible name. r=Jamie
Determine if an accessible is a menuitem within the XUL menu. We know each menuitem is accessible, but it disappears as soon as it is clicked during tests and the a11y-checks do not have time to test the label, because the Fluent localization is not yet completed by then. Thus, we need to special case the label check for these controls.

Differential Revision: https://phabricator.services.mozilla.com/D200101

UltraBlame original commit: 1f8c50293efcae26c2d425cc114d76997c1527c9