Bug 1860887 - Fix autospider WASI unified compile r=jandem
Adding ParseRecordObject.cpp to js/src/moz.build changes the unified file that
includes jsdate.cpp so that it also includes WrappedFunctionObject.cpp, which
contains 'using namespace JS;'. This makes several static functions in
jsdate.cpp ambiguous.

jsdate.cpp was not expecting the JS namespace to be in scope, so it is clear
that the ambiguous function references should all be for the static functions.

Differential Revision: https://phabricator.services.mozilla.com/D200350

UltraBlame original commit: c7e519735961a1991fa1b7fd00debf9399945825