Bug 1863435: Remove wasm::StackMapBoolVector. r=rhunt
We currently do the rather strange process of building stack map info
in a vector, then creating a StackMap, then copying the information over.
This is particularly strange now that we store multiple kinds of entries
in our stack maps. We always know the size of the stack map anyway, so
this patch removes the vector and builds the information in the stack
map directly.

Depends on D196637

Differential Revision: https://phabricator.services.mozilla.com/D196798

UltraBlame original commit: fcafc196f2cef3d5c607edaa0f66c96cdf1708d6