Bug 1878884 - Keep RecordedTextureData snapshot around until it actually changes. r=aosmond
Every BorrowSnapshot use would create and destroy a SourceSurfaceCanvasRecording wrapper
that would go away because it was held by a ThreadSafeWeakPtr, causing us to do a readback
for every GetImageData call, especially when there are multiple such GetImageData calls
in a row.

This replaces the ThreadSafeWeakPtr with a RefPtr, and then adds some MarkChanged tracking
so that we can orphan the snapshot when the underlying DrawTargetRecording changes, but
otherwise keep around the snapshot for successive BorrowSnapshot calls.

Differential Revision: https://phabricator.services.mozilla.com/D201164

UltraBlame original commit: 7ae707f4ae6dc958a4f690a92d47662a1aa77a1c