Bug 1877843 - Migrate javascript.gc GVST metrics r=perry.mcmanis,sfink
Due to JS' relationship with Gecko symbols in general and Telemetry in
particular, this is a side-by-side migration where the Telemetry calls remain
in place and the Glean calls are additional.

(Instead of using GIFFT as an all-in-one, as is preferred).

Differential Revision: https://phabricator.services.mozilla.com/D201015

UltraBlame original commit: 5732bd6e1d62c2864cb9b38bd65d74dffc3c6ac9