Bug 1873723 - Run browsertime tests on beta/release with a cron. r=perftest-reviewers,taskgraph-reviewers,kshampur,jcristau DONTBUILD
This patch modifies the weekly release, and daily beta crons to run the browsertime desktop tests. The crons will also be scheduled at a different time now to be more in line with commits that run tests. At the same time, some of the other crons are modified to use a single method for checking if a platform is a desktop platform that can have tests running on it (similar to what we do for android).

Differential Revision: https://phabricator.services.mozilla.com/D201250

UltraBlame original commit: b39a76100861ab07f99a0455718617676b52adb6