Bug 1877124 - Part 3: Lazily link foreground finalized arenas only when needed by iteration r=sfink
Currently we link the foreground finalized arenas in the sorted arena list into
a single list when we yield to the mutator during foreground finalization.
Iteration is only required for the debugger and memory reporting APIs so it
makes sense to skip this until it's actually needed. This also centralises the
state of which zone/kind we are finalizing in the GCRuntime rather than storing
it in each zone.

Differential Revision: https://phabricator.services.mozilla.com/D200322

UltraBlame original commit: 3e7f6daeeaab8e95d6596f4bc8095df6b099e581