Bug 1859026 - Pt 2. Disable a11y_check for one non-user-facing test of #identity-permission-box in Desktop UI. r=mconley
We intentionally exclude this click from the a11y_check, because this test case (as the ones in part 1 patch D190979) is expected to click a toolbar button that may be already hidden with "display:none;". The permissions panel anchor is hidden because the last permission was removed, however we force opening the panel anyways in order to test that the list has been properly emptied.

Differential Revision: https://phabricator.services.mozilla.com/D201615

UltraBlame original commit: eeb5df11b720bcd86388cdb010a5b5d8edc3bdf3