Bug 1866032 - Add fetchpriority test for initially loaded images. r=valentin,necko-reviewers
For now, we put the same expectation as we chose for tests related to
as=image and dynamically loaded images, that is the ones from [1].
However, we may decide to depart from that later.

This is based on a patch from Ziran Sun (zunigalia.com).

[1] https://web.dev/articles/fetch-priority?hl=en#browser_priority_and_fetchpriority

Differential Revision: https://phabricator.services.mozilla.com/D200952

UltraBlame original commit: 4e991e208a51534f4ab72990e453a89a0d117af6