Bug 1879197 - Part 8: Port "return early if port is closed", r=ipc-reviewers,mccr8
This is a port of a chromium change originally reviewed at
https://chromium-review.googlesource.com/c/chromium/src/+/3440005

Original Author: Stephen Röttger <sroettgergoogle.com>

Original Description:

There was a race where we ran into the DCHECK if we marked the port as
closed in another thread.
There's another DCHECK like this in MergePortsInternal, but I believe
those are fine. We mark the ports as closed if we lose the connection to
another node. MergePortsInternal is either called to merge local ports
or on the IO thread, so there shouldn't be a race here.

Depends on D201111

Differential Revision: https://phabricator.services.mozilla.com/D201112

UltraBlame original commit: 3741f0880152fd5c37d441ff23e3bda9caca8e5c