Bug 1878380 - To AudioCallbackDriver add a ChangingDevice audio stream state. r=padenot
The new state is set when we're notified by cubeb (which is implemented only in
the mac backend) that it is changing output device under the hood. While the new
state is set, any audio callbacks coming in will be ignored, since there is no
guarantee that they won't come in after the DeviceChanged callback, but before
the device switch starts. A fallback driver is started, if not already running,
at the same time as setting the ChangingDevice audio stream state. The fallback
driver changes the audio stream state back to Starting in its first data
callback after a five millisecond grace period. Then handover to the audio
driver happens as it normally would: the first data callback from the audio
driver sets the audio stream state to Running, and the next fallback data
callback after that stops the fallback driver and tells the audio driver it can
drive the graph.

Differential Revision: https://phabricator.services.mozilla.com/D201195

UltraBlame original commit: b82c1a74415870e055e58d8707e4efe8b3948ac4