Bug 1880428 - [devtools] Make line highlight background color semi-transparent so text selection is visible. r=ochameau,devtools-reviewers.
The background color of the line highlight was opaque, hiding any
text selection that could be on that line.
Using a color with some transparency makes the selection visible.
We take this as an opportunity to fix and refactor the CSS for
line highlight.

Differential Revision: https://phabricator.services.mozilla.com/D201941

UltraBlame original commit: 5e0067938b95388f1aa74116463550ecebb40b73