Bug 1879259 - Use GlobalTeardownObserver for consistency on workers. r=dom-storage-reviewers,janv
This removes the manual global tracking which was only correctly hooked
up for windows, but missed other global types, especially workers.
Because GlobalTeardownObserver is reliably notified of when to
disconnect, it is able to "safely" hold a raw non-owning pointer which
simplifies cycle collection somewhat.  (That design decision was
originally part of DOMEventTargetHelper which GlobalTeardownObserver
was extracted from.)  It also provides a pre-cached window getter.

This test includes a WPT test that reliably fails without the fix and
passes with the fix.

Differential Revision: https://phabricator.services.mozilla.com/D201188

UltraBlame original commit: def4a6f94713c0d0e4b2808d42dd486520c00a60