Bug 1877795 [wpt PR 44328] - Add WPT test validating that drawImage writes to the layer, a=testonly
Automatic update from web-platform-tests
Add WPT test validating that drawImage writes to the layer

Added a WPT to check that using `drawImage` inside a layer writes the
image to the layer and not it's parent directly. The test uses
`drawImage` with another canvas as source, inside a layer that applies a
filter and uses the global composite operation `xor`. In its reference,
the same drawing commands are used in the main canvas inside the layer,
to verify that the filter effect is applied on the image, and that the
compositing of the `drawImage` operation is done with the layer's content
and not the parent's.

Bug: 1396372
Change-Id: I377b4eb0b77865c639b296df70110c2fa8f7f07f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253602
Reviewed-by: Fernando Serboncini <fserbchromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravelchromium.org>
Commit-Queue: Andres Ricardo Perez <andresrperezchromium.org>
Cr-Commit-Position: refs/heads/main{#1260079}

--

wpt-commits: fe09a18307f1076a52ad8a94eba004eb939f6ab9
wpt-pr: 44328

UltraBlame original commit: 5acfdf891d03665d6798d9d96b1974f730d099dd