Bug 1874040 - Move 1103348-1.html to WPT, and expand it. r=smaug
This test caught a subtle bug in one of the versions of my previous
patch, but it caught it by chance.

The reason it got caught is because the reftest harness uses a mutation
observer, and thus disables the textContent optimization that reuses the
text node.

WPT didn't catch this, so move the test there, and add tests for the
characterData mutation and removal-and-append explicitly, to not depend
on textContent optimizations.

Differential Revision: https://phabricator.services.mozilla.com/D202230

UltraBlame original commit: 4ce4c83e0d85795f8566bedccf327bd780bad30f