Bug 1648499 - [devtools] Avoid attaching BrowsingContext thread actors in the context of the browser toolbox. r=devtools-reviewers,nchevobbe
In the Browser Console/Toolbox setup, we spawn both Content Process Targets
and WindowGlobal Targets for document running in the same process as the target.

While the debugger will only use the Content process targets and ignore the WindowGlobal ones,
the inspector will only consider the WindowGlobal ones.

Because of this mixed up setup, we have to ensure disabling (avoid attaching)
the thread actor of the WindowGlobal target in the context of the browser toolbox.

Differential Revision: https://phabricator.services.mozilla.com/D167964

UltraBlame original commit: ebcd4a3522060972584f0736af33974ecf448ddb