Bug 1880930 - Remove irrelevant comment from old-configure.in. r=firefox-build-system-reviewers,sergesanspaille
The comment was added in bug 407459, but the code it was relevant to was
moved without the comment in bug 1257448. At the same time, while the
code that came with the comment was moved and is still relevant, the
comment itself refers to code that was removed in bug 1364358.

Differential Revision: https://phabricator.services.mozilla.com/D202189

UltraBlame original commit: 441bee323dbdeee3d6f2eb5e3fe4ea5d5ecf049a