Bug 1881367 part 1: Upgrade some DocumentTimeline asserts to release-asserts, to ensure that we unregister from the refresh driver properly. r=hiro
If these assertions fail, then we're losing track of a refresh driver
registration, which means we're likely to crash (with a non-specific
backtrace). So it's better to intentionally check this condition up-front and
crash in a clearer controlled way if it fails.

Differential Revision: https://phabricator.services.mozilla.com/D202358

UltraBlame original commit: ba7061542711f062815acf2a44604718e97f4570