Bug 1881367 part 2: Upgrade a SMILAnimationController assert to a release-assert, to ensure that we unregister from the refresh driver properly. r=hiro
If this assertion fails, then we're losing track of a refresh driver
registration, which means we're likely to crash (with a non-specific
backtrace). So it's better to intentionally check this condition up-front and
crash in a clearer controlled way if it fails.

Differential Revision: https://phabricator.services.mozilla.com/D202359

UltraBlame original commit: c5a08d136c837992b07223b6ccdb4df6072c22ef