Bug 1881221 - Make it easier to disable the gecko profiler for testing r=profiler-reviewers,aabh
Previously we had to patch this moz.configure file to be able to disable the
gecko profiler to test the tier 3 platforms or the platforms that don't support
the gecko profiler. It wasn't super straightforward to do it and this patch
makes it easier to do by adding a `--disable-gecko-profiler` build option.

To be able to use it, append the following option to your mozconfig file:
```
ac_add_options --disable-gecko-profiler
```

Differential Revision: https://phabricator.services.mozilla.com/D202301

UltraBlame original commit: e074bae30fcb342cf3f62ba19c47201807f167bb