Bug 1876841 - Allow overriding all post-startup-delay settings with a CLI option. r=aglavic,perftest-reviewers
This patch fixes an issue where we weren't able to set the post-startup-delay using a user-supplied option. After this patch, all defaults will be overriden by a user-supplied option and we will no longer force the delay to a maximum of the default. At the same time, the option is adding to our mozharness code so we can set it in our CI tasks.

Differential Revision: https://phabricator.services.mozilla.com/D202638

UltraBlame original commit: 44a27f2fec8088955bee59ca128bea827b02269e