Bug 1882378 - Add an accessible name for Edit pref value inputs on about:config page. r=settings-reviewers,Gijs
When a value is being edited on the about:config page, the second column cell gets a form included with an `<input>` to edit the value of the pref, but this form field is not labeled and users of assistive technology would not know what is the purpose of this form and this input field. This would especially affect screen reader users with cognitive difficulties, memory loss, and neurodivergent users.

Example of the input would include: [this edit input](https://searchfox.org/mozilla-central/rev/9cd4ea81e27db6b767f1d9bbbcf47da238dd64fa/toolkit/components/aboutconfig/content/aboutconfig.js#256-265) and an example of failed a11y_checks tests that would need to have `fail-if` notation to be edited or removed from their test manifests, when this bug is fixed, would be: `toolkit/components/aboutconfig/test/browser/browser_clipboard.js`.

Note that only text and number inputs are being added and are requiring the label, the boolean value is changed with `Toggle` button.

Differential Revision: https://phabricator.services.mozilla.com/D202895

UltraBlame original commit: ea3c7406e68d3b49ba894faf61eb59a32a0427a7