Bug 1880928 - Update some "last remembered sizes" before determining proximity to the viewport. r=emilio,layout-reviewers
We do it right after layout to make sure sizes are up-to-date and before
determining proximities to viewport of 'content-visibility: auto' to
make sure we are not incorrectly recording the size of a previously
"skipping its content" node.

This change is covered by tests added in D202403.

Differential Revision: https://phabricator.services.mozilla.com/D202731

UltraBlame original commit: e01ba93d881ffcc289f284195f013c28f11e158d