Bug 1879121 [wpt PR 44443] - Add permissions .well-known endpoint for Protected Audience API WPTs., a=testonly
Automatic update from web-platform-tests
Add permissions .well-known endpoint for Protected Audience API WPTs.

This .well-known is fetched at the origin of the interest group's owner,
and specifies as a URL parameter the origin of frame that's attempting
to join or leave that interest group.

This is implemented such that the origin of the frame is ignored
altogether, and the determination of which operations are allowed
depends strictly on the origin of the interest group owner, and
specifically on the subdomain of the origin of the interest group owner.
wptserve serves each of its two domains at both the raw domain and each
of five subdomains.

- www: disallows both join and leave
- www1: allows join, but not leave
- www2: allows leave, but not join
- 天気の良い日 / élève: allow both join and leave
- anything else (including no subdomain): returns a 404

More information about the permissions endpoint can be found at:
https://github.com/WICG/turtledove/blob/main/FLEDGE.md#13-permission-delegation

This change depends on
https://github.com/web-platform-tests/wpt/pull/44398
to be usable from web platform tests.

Change-Id: I34f71c2576183856a770dd1a2307f5502941d219
Bug: 40275797
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5277476
Commit-Queue: Orr Bernstein <orrbgoogle.com>
Reviewed-by: mmenke <mmenkechromium.org>
Cr-Commit-Position: refs/heads/main{#1265173}

--

wpt-commits: 2b9d21410646401cd42e830f481f0d8b3ba18639
wpt-pr: 44443

UltraBlame original commit: 344537bd7a9a28941466fe8210330f880b39f046