Bug 1483828 - [Part 2] Fix the expected behaviour of child_focus_frame.html for window_focus.xhtml r=smaug
The changes relate to `t13` is because `t13` is the <html> element,
which is not getting the focus anymore with D198436.

Another change relate to focus_frameset.html is because focus_frameset.html
contains some frames and each frame has a focusable
element in it. With the change made in D198436, these frames are no
longer getting the focus, so the expected behaviour needs to be updated.

Differential Revision: https://phabricator.services.mozilla.com/D200863

UltraBlame original commit: 58b071bd0163d3b0743a64dd32372b48c1db352e