Bug 1483828 - [Part 4] Ensure KEY_F6 still works properly for chrome document navigation r=smaug
KEY_F6 has a special requirement for chrome document navigation
such that if the focus algorithm has reached to the end of the
top-level chrome document, it expects to focus first element of
this chrome document, because the root element of chrome document
is not focusable. This patch is to ensure this behaviour is still
working.

dom/tests/mochitest/chrome/test_focus_docnav.xhtml is used to test this.

Differential Revision: https://phabricator.services.mozilla.com/D200864

UltraBlame original commit: 9f1177bf84f2c131001fb10c8eff671dc56f6383