Bug 1882817 - Export stretchy-largeop-* reftests as WPT testharness.js tests. r=emilio
These tests verify operator stretching for various stretchy or largeop
operators, by checking the `BoundingClientRect` of `<mo>` operators.
This is equivalent to existing more exhaustive checks that are done in
WPT, the difference is that here they are only relying on the fonts
pre-installed on the system. So export these tests assuming they are
checking some basic support with the default math font. Gecko is able
to use scale transforms to emulate operator stretching, so it always
pass the tests even when the system does not provide good math fonts.

Additionally, these tests are written as reftests, but are really only
doing some JavaScript checks, so rewrite them as testharness.js tests.

Differential Revision: https://phabricator.services.mozilla.com/D203399

UltraBlame original commit: 5bb118aeb210e4468237700de28b116a5e96443a