Bug 1883802 - Make `ShadowRoot` notify mutation observers of "parent chain changed" r=smaug
Currently, `ShadowRoot::Unbind` does not notify the observers of the
notification.

With this change, `nsRange` starts removing itself from all `Selection`s if its
root becomes disconnected.  This behavior is different from Chrome, Chrome
allows the disconnected range in shadow as a selection range.  However, it
does not match with the behavior tested in
`move-selection-range-into-different-root.tentative.html` [1][2].  Therefore, I
think that removing the range from `Selection` is reasonable for now because
it's not reasonable to add a check to all selection range getters/users to check
whether the range is in the composed tree.

1. https://searchfox.org/mozilla-central/rev/b60cb73160843adb5a5a3ec8058e75a69b46acf7/testing/web-platform/tests/selection/move-selection-range-into-different-root.tentative.html
2. https://wpt.fyi/results/selection/move-selection-range-into-different-root.tentative.html?run_id=5070898326929408&run_id=5147803004698624&run_id=5098392828510208&run_id=5175063246012416

Differential Revision: https://phabricator.services.mozilla.com/D204218

UltraBlame original commit: dda1e753390ebd7e848058e896da8b74bb105068