Bug 1883909 - P3. Workaround to support detect page navigation for the same-origin iframe. r=credential-management-reviewers,jneuberger,sgalich
When a page navigation event is received in the top-level frame, the same-origin iframe
is able to know that because they are in the same process.

However, the plan is to remove `FormAutofillContent`, which means we
will not be able to use "per process" module to communicate directly.

So instead, we should solve the issue by the child talks to the parent,
and then the parent talks to its childern. But before implementing the
solution mentioned above, we still use `FormAutofillContent` to workaround it.

Differential Revision: https://phabricator.services.mozilla.com/D203767

UltraBlame original commit: b7875a84aa37d61bb72600510063b156861ad115