Bug 1858048 - Part 2: Improve embedded control name computation, r=Jamie
This revision changes Gecko's embedded control name computation to hew to the
relevant piece of the name computation spec (currently part "C" as of writing
this, labelled "Embedded Control"). To accomplish this, this revision moves the
call to AppendFromValue ahead of the call to Name. We use AppendFromValue to do
the embedded control computation, so it's appropriate to place it before we
check aria-label and similar. LocalAccessible::Value (and similar) contain the
necessary means to calculate the things we need, except for in the case of
listbox. So, this revision contains a carveout for listbox. Finally, this change
removes the previously-marked expected-failures from the relevant WPT meta file.

Differential Revision: https://phabricator.services.mozilla.com/D204459

UltraBlame original commit: 334938a86d4160e648fc0d74e6b29adafd133d07