Bug 1884000 - Use ffmpeg instead of windowrecorder for osx 10.15 raptor tests. r=perftest-reviewers,sparky
There is a high failure rate in OSX 10.15 raptor pageload tests. This
patch replaces the windowrecorder with ffmpeg in an attempt to resolve
this as the failures seem to be occuring on the first iteration of when
window recorder is being used.

Differential Revision: https://phabricator.services.mozilla.com/D203934

UltraBlame original commit: 65df374506551b71fd3759cdb913ccc38d3900ea