Bug 1885376: Allow first scene flush to be async. r=nical
This recovers most of the performance degradation of ts_paint, which
measures time to first MozAfterPaint event received. It's not critical
that the first flush be sync, so let's keep things flowing in the way
that ts_paint prefers.

Differential Revision: https://phabricator.services.mozilla.com/D204715

UltraBlame original commit: b0450657a485cfbd9253f14802ce8a3f08503e58