[components] Closes https://github.com/mozilla-mobile/android-components/issues/911 - ui-autocomplete: Do not restart input method in deleteSurroundingText.
This fix is speculative. There seems to be theoretical reason we need
to restart the input method and I think it's creating a race condition
that causes the keyboard to stop working on some devices. In practice,
removing this call fixes the bug but otherwise doesn't seem to make a
difference in autocomplete behavior on device or in the emulator.

UltraBlame original commit: 0d5d662d02315981490dc4d1a85d415d90117eb4