[components] Clear out account state and attempt recovery during authentication errors
This calls a new FirefoxAccount.onAuthTokenError method after encountering
auth errors, which clears out internal token caches in the fxalib.

Right after, we attempt to recover. It's possible our access token simply
expired. We do so by trying to obtain a new access token. If that fails,
we've certainly hit an authentication problem and need user intervention
to recover. Otherwise, we proceed back into authenticated states.

Finally, if we are in trouble and hit a real authentication problem,
clear out any and all mention of the old account data. Currently
there's a caching bug which means we'll re-use a bad access token
during a device refresh even after re-authenticating, and this works
around that bug.

Additionally, this sneaks in one more change: requesting a "profile"
scope is no longer necessary from the application's point of view.
This change is backwards compatible, since passed scopes array
is treated as a set.

This change was necessary to future-proof our token recovery flow.
With it in place, we can be certain that we'll always have the scope
with which we're trying to recover.

UltraBlame original commit: c5c8799eb26e13724c7ac9c52259b8ae773df58e