[components] Merge https://github.com/mozilla-mobile/android-components/pull/4511
4511: Always set 'wasMigrated' to false in the Glean SDK r=Dexterp37 a=Dexterp37

This is only useful as a safety measure in case the data migration plan goes wrong and we're forced to downgrade the version Glean SDK consumers are using from glean-core to glean-ac. In that case, this code will run and will reset the status of the migration. When attempting the upgrade again, glean-core will perform the migration one more time.




Co-authored-by: Alessio Placitelli <alessio.placitelligmail.com>

UltraBlame original commit: c19b8cde49ffc4a1a2849511fd515bf262ac05db