[components] Merge https://github.com/mozilla-mobile/android-components/pull/4762 https://github.com/mozilla-mobile/android-components/pull/4769 (part 1)
4762: For https://github.com/mozilla-mobile/android-components/issues/4761 - Adds setting to control `forceUserScalableEnabled` r=csadilek a=boek



4769: For https://github.com/mozilla-mobile/android-components/issues/2923: Don't crash if Intl.RelativeTimeFormat not supported r=Amejia481 a=csadilek

`Intl.RelativeTimeFormat` is in Nightly and Beta, but not yet on Release (until Monday). Not a problem as Fenix is on Beta (and R-B on Nightly) anyway, but good to catch errors anyway. This would also help if an article specified an invalid lang for instance.

Co-authored-by: Jeff Boek <jeffjeffboek.com>
Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 20c23d4110139f2f2a804980c5a6627bffe68517