[components] Merge https://github.com/mozilla-mobile/android-components/pull/4977
4977: Closes https://github.com/mozilla-mobile/android-components/issues/4969: Provide API to observe browser state changes on all tabs r=pocmo a=csadilek

This allows us to map a flow to a list (e.g. to a list of tabs as in `state.tabs`) and `collect` individual elements for which a specific attribute has changed (e.g. only tabs for which `windowRequest` has changed).

pocmo this is pretty much as discussed, relying on the new `flatMap` flow operators. I did verify that this works and allows me to get rid of the workaround, but curious what you think. I also need to sleep over this :). In general, I think it does make sense that we have a `ifChanged` for flows of states, and `filterChanged` for flows of list of states. 

Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 8e9e1c208db0ae621302f6141453dd4bc3e63a0f