[components] Closes https://github.com/mozilla-mobile/android-components/issues/5063: don't log unneccessary illegal state exceptions
We are logging exceptions when no profile is found, which is actually a legit case.
There will be no profile on a fresh install, for example.

This patch makes these kinds of exceptions more precise/useful: we'll only send a report
in case there's a profile but a missing dbPath. We'll now also send an exception in case
of an IOException while parsing profiles.

UltraBlame original commit: 0859adf7777e7afdaa968d507f5cb407ab0abf1d