[components] Merge https://github.com/mozilla-mobile/android-components/pull/4985
4985: Create wrapper for Glean gradle plugin r=mdboom a=mdboom

Glean now has a Gradle plugin for generating metrics from metrics.yaml files.

Since this is built and published alongside Glean itself, we need a way to load the plugin so that the version of the plugin matches the version of Glean being used.  Unfortunately, most of our users are using Glean *through* android-components, so they don't know the Glean version.  This wrapper plugin will allow it to be loaded using the android-components version, and the publishing will happen on the same cadence as android-components itself.



Co-authored-by: Michael Droettboom <mdboomgmail.com>

UltraBlame original commit: b7c240f9dc5513cdffd3cf923b697875754f9c0d