[components] Merge https://github.com/mozilla-mobile/android-components/pull/5526
5526: Closes https://github.com/mozilla-mobile/android-components/issues/5524: Intermittent failure in WebExtensionSupportTest r=Amejia481 a=csadilek

This is another example where we need our new `waitUntilIdle` as a method we're calling dispatches an action that we can't join on in the test. So, we have to wait until the store processed the action. Looked over the other tests in this class as well. This should be the only place that's needed right now.

Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 3b57e5d1d0cc2ca798e33981108537bb2d55dfac