[components] Merge https://github.com/mozilla-mobile/android-components/pull/5612
5612: Closes https://github.com/mozilla-mobile/android-components/issues/5588: WebExtensionToolbarFeature may display stale actions r=psymoon,Amejia481 a=csadilek

This makes sure we remove stale browser actions (from uninstalled extensions) from the toolbar. Reason we don't always see this bug is that when switching from the `AddonsFragment` (where we uninstall) back to the `BrowserFragment` we usually get a new view and toolbar and we'd only add the actions of existing extensions. However, when the view and toolbar are re-used we'd see this problem (easily reproducible in an emulator).

Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: aeb11c3b4d8869fc653beed518db608aa507705c