[components] Merge https://github.com/mozilla-mobile/android-components/pull/6151
6151: Closes https://github.com/mozilla-mobile/android-components/issues/6018: Do not load url during link if session is restored r=pocmo,csadilek a=psymoon

The potential pitfall of this patch is we are assuming `geckoSession.restoreState()` would always be successful. Perhaps we could have a follow-up for GV to provide us a callback for when `restoreState()` is completed with a status? 🤔 

Co-authored-by: Simon Chae <chaesmngmail.com>

UltraBlame original commit: 9888a7ea055ab3830d749e93d3914b52a86ae10d