[components] Merge https://github.com/mozilla-mobile/android-components/pull/6843
6843: Closes https://github.com/mozilla-mobile/android-components/issues/2985: Launch reader view in extension page r=jonalmeida a=csadilek

Design is as described in https://github.com/mozilla-mobile/android-components/issues/2985. Instead of transforming the current DOM, we now open reader view in an extension page. We inject only a small script (instead of the whole logic) into every page to check whether or not the page is readerable. If the reader view is activated, we send a message to a background script to open the reader page which contains the rest of the logic and its dependencies.

This still requires a few workarounds, which I documented with links to the corresponding issues. However, this is now finally in a functioning and landable state. It solves many problems users reported in Fenix e.g. reader stuck in wrong zoom level, reader refreshing or reloading the original page, overlays/popups appearing etc. Also, it is now possible to navigate back and forth into reader view as we have a history entry (the extension page).

Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 34632d106733b0c0b705d7169e0cec35e92738ed