[components] Merge https://github.com/mozilla-mobile/android-components/pull/6902 https://github.com/mozilla-mobile/android-components/pull/6906 (part 2)
6902: Closes https://github.com/mozilla-mobile/android-components/issues/6881: Register/unregister download notification actions receiver r=csadilek a=Amejia481



6906: Issue https://github.com/mozilla-mobile/android-components/issues/2985: Support loading reader view ext. page in new session r=jonalmeida a=csadilek

Basically, a leftover from the refactoring. We used to turn on reader mode when the port connected and reader mode was active for this session. Now we have a separate page which could be loaded in a new session. When the reader extension page connect, we can simply "show" reader mode (and mark it as active for the session, but that part already worked.)

Co-authored-by: Arturo Mejia <arturomejiamarmolgmail.com>
Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: ff5314fdf9f303a45a59a3cd90ed94976cc18e3b