[components] For https://github.com/mozilla-mobile/android-components/issues/7627 - Add hasRegionCached() to LocationService
This method can be used by clients to better estimate if fetchRegion() is
expected to take very little time as the region is already cached or if it can
take a bit longer since there are multiple operations / longer running
operations needed to actually fetch the region.

In the case of MozillaLocationService if the region is not already cached the
process of fetching the region involves a REST call with a connect / read
timeout of 10 seconds.

UltraBlame original commit: b21e6fde6e6bf85065abbbfab3ecf2bb4f87c8fb