[components] For https://github.com/mozilla-mobile/android-components/issues/7992 - Don't show "Download link" for webpages links
The "Download link" option in the contextual menu might not be a necessary
option if the link is for other webpages.
Speculate based on the url ending whether the link is for a webpage and prevent
showing the "Download link" in this case.

UltraBlame original commit: 17f0959557c7238bf4f87062e9897f3ee871b038