[components] For https://github.com/mozilla-mobile/android-components/issues/8356 - Speculative fix for BrowserGestureDetector IAE
Try to resolve a situation in which system's GestureDetector calls us as a
callback with null parameters, so triggering an IllegalArgumentException.
A possible explanation could be that not all MotionEvents are passed to the
GestureDetector, situation which this patch tries to fix.

UltraBlame original commit: 3fba2fb7970415cce09210f831571aced10cba01