[components] Merge https://github.com/mozilla-mobile/android-components/pull/8358
8358: SessionStorage: Write state as stream to disk instead of building model in memory first. r=Amejia481 a=pocmo

With this change we are using `JsonWriter` to write `BrowserState` directly to disk using a stream. Previously we first build a `JSONObject` tree in memory. This essentially duplicates the state and that could become big if a user had a lot of tabs open and not much RAM.

Co-authored-by: Sebastian Kaspari <s.kasparigmail.com>

UltraBlame original commit: 46761820cbf8f86d872daf784cbbda49ab31cf28