[components] Don't run tasks for mergify branches on push
This is duplicate, unnecessary work. We should do one or the other, and
I _think_ it makes more sense to use the PR ones (that's what we do for
every other kind of PRs). The only thing that makes these special vs.
non-mergify PRs is that the branches happen to be in this repository -
so taskcluster notices the branches.

UltraBlame original commit: 056fb533a1f80543964b85e2b9eaef6df48f4073