[components] For https://github.com/mozilla-mobile/android-components/issues/10005 - Replace layout manager to effectively change stackFromEnd
The `stackFromEnd` property is not initially set even though
`endOfMenuAlwaysVisible` might be true since when having a bottom expandable
menu, that should always be scrolled to the top.
And we know if we're in that situation only later, after menu is configured and
based on it it can be inferred if it should be collapsed or not.

Setting `stackFromEnd` only after this steps, even if the menu is not yet on
the screen seems to have no effect.

The only solution is to replace menu's layout manager with one that has this
`stackFromEnd` from the beginning.

UltraBlame original commit: 41abfdb425ff89cd27e349a66bfd6b9d096abdb2