[components] Updating browser-engine-gecko to work with the new logins API
Switched the login saving interfaces to using `LoginEntry`, which
matches the change in `LoginValidationDelegate`.

The select login code still uses `Login`, but I changed how the
conversion from `GeckoView.LoginEntry` works.  Now it checks if the guid
is set, and ignores the login if not.  This is needed because a) guid is
no longer nullable and b) `areItemsTheSame()` in `BasicLoginAdapter.kt`
uses the guid to compare rows in the select list.  I'm pretty sure the
guid should always be set because those login entries ultimately come
from a `List<Login>`.

Removed the `GeckoView.LoginEntry` -> `Login` conversion from
`ext/Login.kt`.  It's now more dangerous to call, since it will throw if
guid is null.  I think it's better to force the calling code to write
the `!!` themselves to make this clear.

UltraBlame original commit: 342cc94751be90ca23770fde5dfa26f7e59093a0