[components] Handle SCOPE_SYNC access tokens missing a 'key'
For reasons that are currently vague and not understood well, we could
be getting SCOPE_SYNC access tokens without a 'key'. Currently, that
will cause a crash. This patch adds just enough handling of this state
to avoid crashing, transition the account into one of the terminal
states ('not authenticated' if this happened during a login, 'auth
problem' if this happened afterwards), and report an 'info' via the
crash reporter with enough information to understand when exactly this
happens.

UltraBlame original commit: fb9246b8ac259d15ab814430bc7f9248c876c0a8