[focus] Set severity of lint issues MissingTranslation and ExtraTranslation to warning.
When adding or removing strings the app build fails because some strings are
not translated or do not exist in strings.xml anymore. This is expected until
we do a new import. So let's not fail builds in this case.

UltraBlame original commit: bec25021a590d8742236400eddc73f10740d7f9b