[focus] Add static build of telemetry library (https://github.com/mozilla-mobile/focus-android/issues/18).
This patch adds a static AAR of the current state of the telemetry library [1]. This is not
optimal and in the future we should deploy the library to a maven repository. Either something
like jcentral or host our own one. For now I'll update the library manually.

[1] https://github.com/mozilla-mobile/telemetry-android

UltraBlame original commit: 29f7b608843db22fce09f32a5eaa3ae942b48083