[focus] Migrate UrlMatcher.matches() to use Uri
We used to convert the Uri to a String, only to parse it as URL
in order to call getHost() on it. We can instead just pass the original
Uri around.

(This is a historical remainder from when we used to support the older
 shoudlInterceptRequest() version, which would only give us the resource
 URL string, hence UrlMatcher was designed around handling a URL string.)

UltraBlame original commit: 76620011d93f72b27cc79e5400131111639dbf87