[focus] Issue https://github.com/mozilla-mobile/focus-android/issues/1936: normalize user-entered search engines.
In particular, this will add http:// if it was missing, which is a common error
because it is missing in the example.

Alternatively, we could validate that the user includes a protocol in the entry
screen but given we're in a rush, I didn't get UX feedback and we should never
really be sending a user to about:blank because they don't have a protocol
anyway so I decided to make this change.

UltraBlame original commit: 993fb9f8cae36445d8b3bd02e78472b909a6e32b