[focus] For https://github.com/mozilla-mobile/focus-android/issues/4975 - Add local.properties autosignReleaseWithDebugKey
Using a new local.properties file we can declare various properties that will
be used to configure builds.
This file can contain secret keys or various settings only used for local
development so it must not be committed.

The first property that can be used from this file is
autosignReleaseWithDebugKey.

It only needs to be declared in local.properties and then all builds, including
the release builds will automatically be signed with the debug key easing
development and testing.

As an edgecase for the Focus release build if an Adjust token is not provided
but autosignReleaseWithDebugKey is declared then a fake one will be used just
to have the app run normally. The Adjust functionality will be broken but a
release build without an Adjust token would otherwise crash at startup.

UltraBlame original commit: 27d816d2be6fdcbf6c659e379af29719e350642b