[focus] FOr https://github.com/mozilla-mobile/focus-android/issues/5133: Changed code path for first run
The default code path used to be to default to the First Run screen. However, the app could default to Home and switched to the First Run screen
when the app is first run. This also helps when running other experiments / tests (in my case it would be performance tests)

UltraBlame original commit: 53d94824af881a0e7d87f505ae48690559716144